docs+adjust: Day 3–4 前语义锁定——handoff、ADR 修订、protocol/schema 调整、roadmap 重排

+++ packages/protocol/src/report.ts
@@ -68,7 +68,7 @@ export const reportRequestSchema = z.object({
   finalText: z.string().optional(),
   /** Workflow cursor (free-form) → persisted as loop.state for the next run. */
   cursor: z.unknown().optional(),
-  durationMs: z.number().nonnegative().optional(),
+  durationMs: z.number().int().nonnegative().optional(),
   /** Agent session id on the machine (locates the local transcript). */
   sessionId: z.string().optional(),
   /** Latest content of the loop's task file (durable context+log doc). */
